Trait sui_indexer_alt::handlers::Handler

pub trait Handler {
    type Value: Send + Sync + 'static;

    const NAME: &'static str;
    const FANOUT: usize = 10usize;
    const BATCH_SIZE: usize = 50usize;
    const CHUNK_SIZE: usize = 200usize;
    const MAX_PENDING_SIZE: usize = 1_000usize;

    // Required methods
    fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>>;
    fn commit<'life0, 'life1, 'life2, 'async_trait>(
        values: &'life0 [Self::Value],
        conn: &'life1 mut Connection<'life2>,
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Handlers implement the logic for a given indexing pipeline: How to process checkpoint data into rows for their table, and how to write those rows to the database.

The handler is also responsible for tuning the various parameters of the pipeline (provided as associated values). Reasonable defaults have been chosen to balance concurrency with memory usage, but each handle may choose to override these defaults, e.g.

Handlers that produce many small rows may wish to increase their batch/chunk/max-pending sizes).

Handlers that do more work during processing may wish to increase their fanout so more of it can be done concurrently, to preserve throughput.

Required Associated Types

type Value: Send + Sync + 'static

The type of value being inserted by the handler.

Required Associated Constants

const NAME: &'static str

Used to identify the pipeline in logs and metrics.

Provided Associated Constants

const FANOUT: usize = 10usize

How much concurrency to use when processing checkpoint data.

const BATCH_SIZE: usize = 50usize

If at least this many rows are pending, the committer will commit them eagerly.

const CHUNK_SIZE: usize = 200usize

If there are more than this many rows pending, the committer will only commit this many in one operation.

const MAX_PENDING_SIZE: usize = 1_000usize

If there are more than this many rows pending, the committer applies backpressure.

Required Methods

fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>>

The processing logic for turning a checkpoint into rows of the table.

fn commit<'life0, 'life1, 'life2, 'async_trait>( values: &'life0 [Self::Value], conn: &'life1 mut Connection<'life2>, ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Take a chunk of values and commit them to the database, returning the number of rows affected.

Object Safety

This trait is not object safe.