Function sui_indexer_alt::pipeline::concurrent::pipeline

pub fn pipeline<H: Handler + 'static>(
    initial_watermark: Option<CommitterWatermark<'static>>,
    config: PipelineConfig,
    db: Db,
    checkpoint_rx: Receiver<Arc<CheckpointData>>,
    metrics: Arc<IndexerMetrics>,
    cancel: CancellationToken,
) -> (JoinHandle<()>, JoinHandle<()>, JoinHandle<()>)

Start a new concurrent (out-of-order) indexing pipeline served by the handler, H. Starting strictly after the watermark (or from the beginning if no watermark was provided).

Each pipeline consists of a processor task which takes checkpoint data and breaks it down into rows, ready for insertion, and a committer which writes those rows out to the database. Committing is performed out-of-order: the pipeline may write out checkpoints out-of-order, either because it received the checkpoints out-of-order or because of variance in processing time.

The committer also maintains a row in the watermarks table for the pipeline which tracks the watermark below which all data has been committed (modulo pruning).

Checkpoint data is fed into the pipeline through the checkpoint_rx channel, and an internal channel is created to communicate checkpoint-wise data to the committer. The pipeline can be shutdown using its cancel token.