Trait sui_indexer_alt::pipeline::concurrent::Handler

pub trait Handler: Processor {
    const MIN_EAGER_ROWS: usize = 50usize;
    const MAX_CHUNK_ROWS: usize = 1_000usize;
    const MAX_PENDING_ROWS: usize = 5_000usize;
    const WRITE_CONCURRENCY_OVERRIDE: Option<usize> = None;

    // Required method
    fn commit<'life0, 'life1, 'life2, 'async_trait>(
        values: &'life0 [Self::Value],
        conn: &'life1 mut Connection<'life2>,
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided method
    fn prune<'life0, 'life1, 'async_trait>(
        _from: u64,
        _to: u64,
        _conn: &'life0 mut Connection<'life1>,
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Handlers implement the logic for a given indexing pipeline: How to process checkpoint data (by implementing Processor) into rows for their table, and how to write those rows to the database.

The handler is also responsible for tuning the various parameters of the pipeline (provided as associated values). Reasonable defaults have been chosen to balance concurrency with memory usage, but each handle may choose to override these defaults, e.g.

Handlers that produce many small rows may wish to increase their batch/chunk/max-pending sizes).

Handlers that do more work during processing may wish to increase their fanout so more of it can be done concurrently, to preserve throughput.

Concurrent handlers can only be used in concurrent pipelines, where checkpoint data is processed and committed out-of-order and a watermark table is kept up-to-date with the latest checkpoint below which all data has been committed.

Back-pressure is handled through the MAX_PENDING_SIZE constant – if more than this many rows build up, the collector will stop accepting new checkpoints, which will eventually propagate back to the ingestion service.

Provided Associated Constants

const MIN_EAGER_ROWS: usize = 50usize

If at least this many rows are pending, the committer will commit them eagerly.

const MAX_CHUNK_ROWS: usize = 1_000usize

If there are more than this many rows pending, the committer will only commit this many in one operation.

const MAX_PENDING_ROWS: usize = 5_000usize

If there are more than this many rows pending, the committer applies backpressure.

const WRITE_CONCURRENCY_OVERRIDE: Option<usize> = None

Provides a way for individual pipeline to override the write_concurrency parameter from the PipelineConfig. This is used to determine the number of concurrent tasks to commit data to the database.

Required Methods

fn commit<'life0, 'life1, 'life2, 'async_trait>( values: &'life0 [Self::Value], conn: &'life1 mut Connection<'life2>, ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Take a chunk of values and commit them to the database, returning the number of rows affected.

Provided Methods

fn prune<'life0, 'life1, 'async_trait>( _from: u64, _to: u64, _conn: &'life0 mut Connection<'life1>, ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait,

Clean up data between checkpoints _from and _to (inclusive) in the database, returning the number of rows affected. This function is optional, and defaults to not pruning at all.

Object Safety

This trait is not object safe.